requirements.txt
numpy==1.23.3
RPi.GPIO==0.7.1
smbus==1.1.post2
